refactor(inference): rename folder
