+++ packages/backend/src/managers/inference/inferenceManager.spec.ts
