+++ packages/backend/src/studio-api-impl.spec.ts
@@ -24,7 +24,7 @@ import userContent from './ai-user-test.json';
 import type { ApplicationManager } from './managers/applicationManager';
 import { StudioApiImpl } from './studio-api-impl';
 import type { PlayGroundManager } from './managers/playground';
-import type { InferenceManager } from './managers/playground/inferenceManager';
+import type { InferenceManager } from './managers/inference/inferenceManager';
 import type { TelemetryLogger, Webview } from '@podman-desktop/api';
 import { EventEmitter } from '@podman-desktop/api';
 import { CatalogManager } from './managers/catalogManager';
